+++ rust-src/concordium_base/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add constructor `TokenAddress::new` for CIS2 type `TokenAddress`.
 - Introduce new chain parameters `ValidatorScoreParameters` that contain the
   threshold of maximal missed rounds before a validator gets suspended.
+- Add `Default` instance for `UpdateSequenceNumber`.
 
 ## 6.0.0