+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased changes
 
+## 7.0.0 (2025-02-03)
+
 - Add getter function `reward_period_epochs` to access the field in the struct `RewardPeriodLength`.
 - Add constructor `TokenAddress::new` for CIS2 type `TokenAddress`.
 - Introduce new chain parameters `ValidatorScoreParameters` that contain the
@@ -7,7 +9,7 @@
 - Add `Default` instance for `UpdateSequenceNumber`.
 - Add `FinalizationCommitteeHash` type.
 
-## 6.0.0
+## 6.0.0 (2024-08-26)
 
 - Extend `id::types::ATTRIBUTE_NAMES` with new company attribute tags: "legalName", "legalCountry", "businessNumber" and "registationAuth".
 - Add a new module `cis3_types` that defines the interface types for CIS3